NOISSUE: move generic parts of slotdebugger to common testutils, introduce virtual slotdebugger with runner for virtual semi integrational tests
