+++ ledger-core/v2/testutils/messagesender/adapter_wrapper.go
