+++ ledger-core/v2/virtual/object/object_test.go
@@ -19,9 +19,9 @@ import (
 	"github.com/insolar/assured-ledger/ledger-core/v2/pulse"
 	"github.com/insolar/assured-ledger/ledger-core/v2/reference"
 	"github.com/insolar/assured-ledger/ledger-core/v2/testutils/gen"
+	"github.com/insolar/assured-ledger/ledger-core/v2/testutils/slotdebugger"
+	"github.com/insolar/assured-ledger/ledger-core/v2/testutils/stepchecker"
 	"github.com/insolar/assured-ledger/ledger-core/v2/vanilla/longbits"
-	"github.com/insolar/assured-ledger/ledger-core/v2/virtual/testutils/slotdebugger"
-	"github.com/insolar/assured-ledger/ledger-core/v2/virtual/testutils/stepchecker"
 )
 
 func Test_Delay(t *testing.T) {
@@ -163,13 +163,7 @@ func TestSMObject_Semi_CheckAwaitDelegateIsStarted(t *testing.T) {
 	smObject.ActiveMutablePendingCount = 1
 
 	slotMachine := slotdebugger.New(ctx, t, true)
-	slotMachine.PrepareMockedMessageSender(mc)
-
-	{
-		pd := pulse.NewFirstPulsarData(10, longbits.Bits256{})
-		pulseSlot := conveyor.NewPresentPulseSlot(nil, pd.AsRange())
-		slotMachine.AddDependency(&pulseSlot)
-	}
+	slotMachine.InitEmptyMessageSender(mc)
 
 	smWrapper := slotMachine.AddStateMachine(ctx, smObject)
